backend/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Security & Auth
# IMPORTANT: bcrypt 4.x has compatibility issues with older passlib
# Use bcrypt 4.0.1 specifically and ensure passlib is updated
bcrypt==4.0.1
passlib[bcrypt]==1.7.4

# JWT
python-jose[cryptography]>=3.3.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Environment
python-dotenv>=1.0.0

# PDF & Document Processing
weasyprint>=60.0
pypdf>=3.0.0
python-docx>=1.1.0
Pillow>=10.0.0

# Template Engine
pystache>=0.6.5

# AI/OpenAI
openai>=1.0.0
httpx>=0.25.0